Refactor setup.py
